% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{browseSignature}
\alias{browseSignature}
\title{Displaying BugSigDB signatures pages in a web browser}
\usage{
browseSignature(sname)
}
\arguments{
\item{sname}{character. Signature name. Expected to start with a prefix
of the form \code{"bsdb:<X>/<Y>/<Z>_"} encoding the corresponding
BugSigDB signature ID.}
}
\value{
The URL of the selected BugSigDB signature page. If interactive,
opens the URL in the default web browser.
}
\description{
Functionality for programmatically displaying microbe 
signatures on BugSigDB signature pages.
}
\examples{

  sname <- "bsdb:215/1/1_eczema:infant-with-eczema_vs_healthy-control_UP"
  browseSignature(sname)
  
}
\references{
BugSigDB: \url{https://bugsigdb.org}
}
