% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPairwiseOverlaps.R
\name{makeDist}
\alias{makeDist}
\title{Title Create a Jaccard or overlap dissimilarity matrix from the
output of calcPairwiseOverlaps}
\usage{
makeDist(overlaps, use = "jaccard")
}
\arguments{
\item{overlaps}{A data.frame output from the calcPairwiseOverlaps function}

\item{use}{Either "overlap" or "jaccard"}
}
\value{
a Dist object of (1 - overlap) or (1 - jaccard)
}
\description{
Title Create a Jaccard or overlap dissimilarity matrix from the
output of calcPairwiseOverlaps
}
\examples{
testlist <- list(a = 1:3, b = 3, c = 3:4)
(all <- calcPairwiseOverlaps(testlist))
makeDist(all, "jaccard")
makeDist(all, "overlap")
}
