% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugsigdb.R
\name{importBugSigDB}
\alias{importBugSigDB}
\title{Obtain published microbial signatures from bugsigdb.org}
\usage{
importBugSigDB(cache = TRUE)
}
\arguments{
\item{cache}{logical. Should a locally cached version used if available?
Defaults to \code{TRUE}.}
}
\value{
a \code{\link{data.frame}}.
}
\description{
Obtain published microbial signatures from bugsigdb.org
}
\examples{

 df <- importBugSigDB()

}
\references{
BugSigDB: \url{https://bugsigdb.org}
}
